another update

+++ c_loop_benchmark.c
@@ -3,15 +3,15 @@
 #include <stdint.h> 
 #include <time.h> 
 
-int length = 2048 * 2048;
+long long length = 2048 * 2048;
 
 struct timespec startTime, endTime;
 
 void PerformArithmetic() {
     clock_gettime(CLOCK_MONOTONIC, &startTime);
 
-    int total = 1;
-    for (int i = 1; i < length; i++) {
+    long long total = 1;
+    for (long long i = 1; i < length; i++) {
         total += i;
     }
 
@@ -19,7 +19,7 @@ void PerformArithmetic() {
     
     double time_taken = (endTime.tv_sec - startTime.tv_sec) +
                     (endTime.tv_nsec - startTime.tv_nsec) / 1e9;
-    printf("Time taken = %.16f seconds, total = %d\n",  time_taken, total);
+    printf("Time taken = %.16f seconds, total = %lld\n",  time_taken, total);
 }
 
 int main() {
